Chore: Move MCP request page into debug router path (#9140)
* refactor mcp page router path
* fix lint issue
