+++ packages/insomnia/src/routes/organization.$organizationId.project.$projectId.workspace.new.tsx
@@ -204,7 +204,7 @@ export async function clientAction({ request, params }: Route.ClientActionArgs) 
       });
 
       return redirect(
-        href('/organization/:organizationId/project/:projectId/workspace/:workspaceId/mcp/request/:requestId', {
+        href('/organization/:organizationId/project/:projectId/workspace/:workspaceId/debug/request/:requestId', {
           organizationId,
           projectId,
           workspaceId: workspace._id,
